Fix a regression in persepctive saving/loading for panes with empty names.
Panes with empty names are meant to have a 'random' name assigned when the paneinfo is added to the frame manager, due to a regression in the code this wasn't happening, which in turn causes issues when saving/loading the perspective.
Issue was only visible if two or more panes with names were present.